Renovate example-host
* Bump dependencies
* Update code to keep up with new dependencies
* Add sync version of calling adder guest
* Some refactoring
